not ignore the api endpoint
actually to deploy things on gh-pages, gh need to know those files.

On a second thought, it's like okay to expose just the endpoint I guess, and no one should be boring enough as to spamming me, :P

However, another solution also exist (which is not too much techy but i just lazy to do it for now but will always be an option) is that, create another repo that is contains those sensitive informations, and add this private as sub-module to the public module.

just too lazy to do for now.